requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
contextvars==2.4
cymem==2.0.6
dataclasses==0.8
datasets==1.16.1
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.3
filelock==3.4.0
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
huggingface-hub==0.2.1
idna==3.3
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
mistune==0.8.4
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
nltk==3.6.5
notebook==6.4.6
numpy==1.19.5
oauthlib==3.1.1
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathy==0.6.1
pickleshare==0.7.5
Pillow==8.4.0
pke @ git+https://github.com/boudinfl/pke.git@f651015f9c931cf245a753f4457bb49f0befa5fd
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydeck==0.6.2
pyDeprecate==0.3.1
Pygments==2.10.0
Pympler==1.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.5.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
streamlit==1.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.10.1+cu102
torchaudio==0.10.1+cu102
torchmetrics==0.6.2
torchvision==0.11.2+cu102
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
transformers==4.14.1
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.2
urllib3==1.26.7
validators==0.18.2
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
